# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TALABAT Multi-Restaurant SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants/{id}/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
        - name: categoryId
          in: query
          schema: { type: number }
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ── Addons ──────────────────────────────────────────────────────────────────
  /products/{id}/addons:
    get:
      operationId: listAddons
      tags: [addons]
      summary: List addons for a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Addons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Addon"

    post:
      operationId: createAddon
      tags: [addons]
      summary: Create addon
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddonInput"
      responses:
        "201":
          description: Addon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Addon"
  /addons/{id}:
    patch:
      operationId: updateAddon
      tags: [addons]
      summary: Update addon
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddonPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Addon"

    delete:
      operationId: deleteAddon
      tags: [addons]
      summary: Delete addon
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ── Drivers ─────────────────────────────────────────────────────────────────
...
  /public/restaurants/{slug}:
    get:
      operationId: getPublicRestaurant
      tags: [public]
      summary: Get restaurant by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Public restaurant info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicRestaurant"
        "404":
          description: Not found
  /public/restaurants/{slug}/menu:
    get:
      operationId: getPublicMenu
      tags: [public]
      summary: Get full menu for customer ordering
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full menu with categories and products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicMenu"
  /public/restaurants/{slug}/orders:
    post:
      operationId: placeOrder
      tags: [public]
      summary: Place a delivery order (guest checkout)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderConfirmation"
        "400":
          description: Restaurant unavailable or invalid data
        "403":
          description: Restaurant subscription expired

  # ── Analytics ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, restaurantId, categoryId, name, price, isAvailable]
      properties:
        id: { type: number }
        restaurantId: { type: number }
        categoryId: { type: number }
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        descriptionAr: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        price: { type: number }
        isAvailable: { type: boolean }
        sortOrder: { type: number }
        addons:
          type: array
          items:
            $ref: "#/components/schemas/Addon"
    ProductInput:
      type: object
      required: [categoryId, name, price]
      properties:
        categoryId: { type: number }
        name: { type: string, minLength: 1 }
        nameAr: { type: string }
        description: { type: string }
        descriptionAr: { type: string }
        imageUrl: { type: string }
        price: { type: number, minimum: 0 }
        isAvailable: { type: boolean }
        sortOrder: { type: number }
    ProductPatch:
      type: object
      properties:
        categoryId: { type: number }
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        descriptionAr: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        price: { type: number }
        isAvailable: { type: boolean }
        sortOrder: { type: number }

    # ── Addon ───────────────────────────────────────────────────────────────────
    Addon:
      type: object
      required: [id, productId, name, price, isAvailable]
      properties:
        id: { type: number }
        productId: { type: number }
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        price: { type: number }
        isAvailable: { type: boolean }
    AddonInput:
      type: object
      required: [name, price]
      properties:
        name: { type: string, minLength: 1 }
        nameAr: { type: string }
        price: { type: number, minimum: 0 }
        isAvailable: { type: boolean }
    AddonPatch:
      type: object
      properties:
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        price: { type: number }
        isAvailable: { type: boolean }

    # ── Driver ──────────────────────────────────────────────────────────────────
...
    OrderConfirmation:
      type: object
      required: [orderId, status]
      properties:
        orderId: { type: number }
        status: { type: string }
        message: { type: string }

    # ── Order Input (public guest checkout) ──────────────────────────────────────
    OrderInput:
      type: object
      required: [customerName, customerPhone, latitude, longitude, items]
      properties:
        customerName: { type: string, minLength: 1 }
        customerPhone: { type: string }
        notes: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: number }
        quantity: { type: number, minimum: 1 }
        selectedAddonIds:
          type: array
          items: { type: number }

    # ── Public storefront ────────────────────────────────────────────────────────
    PublicRestaurant:
      type: object
      required: [id, name, slug, address, status]
      properties:
        id: { type: number }
        name: { type: string }
        slug: { type: string }
        address: { type: string }
        logoUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        primaryColor: { type: ["string", "null"] }
        status: { type: string }
        isAcceptingOrders: { type: boolean }
    PublicMenu:
      type: object
      required: [restaurant, categories]
      properties:
        restaurant:
          $ref: "#/components/schemas/PublicRestaurant"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PublicCategory"
    PublicCategory:
      type: object
      required: [id, name, products]
      properties:
        id: { type: number }
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"

    # ── Analytics ────────────────────────────────────────────────────────────────
...